Issue #7 solved
Program2 is an example of a program where in the bytecode a statement is between a new and an init.

NEW a
f(b)
init a

When we instrument that statement, the call to the dummy method of  InstrumentedMethod includes "a" as a parameter, which wasn't initialized yet.
I fixed that by keeping track of these situations with an additional data structure and not including non initialized variables as parameters