Important!
If a class isn't instrumented - because statically we can't infer that it is used,
for e.g. because the program uses interfaces or reflection,
but dynamically it IS indeed used,
it MUST be visible for daikon.

That's why the non-instrumented classes must be added to classpath
(but with lower priority that the instrumented ones)